Fix preimage gadget test impl
For some reason, use `composer.public_inputs` with `trace` and
`trace-print` features was returning a Public Input vector which was
smaller just by one element in comparaison to the circuit size.

Since we don't use any Public Inputs, seting the vec as zeroes solves
the problem.

Anyway, we should consider to remove a lot of stuff from the
zk::mod.rs file TBH.