Commenting mod.rs
As it stands, the refactor of
phoenix will not require the
logic in the mod.rs file.
Some file functions only cause
build errors which disrupt other
parts of the code.

The fallible code has been
commented out for the meantime
so other parts of the library
can be focused on.