Refactor nullifier gadget
This refactor adjusts the gadget to work on a single TransactionInput,
as seen on a normal transaction (not a zk::ZkTransaction). This rids the need for
reconstructing the secret key on the circuit from the bits of the scalar.

Additionally, it fixes the test by foregoing the use of the lazy_static transcript and parameters.