Remove errors
There are two seperate things done in this commit. The first is changing the slice of bytes to be the public input in the composer for the sum gadget. The second is to remove the warnings; it is worth noting the remaining 2 warnings will be removed when the zk/mod.rs is fully refactored