Add range gadget
The range gadget takes in any type that implements the TransactionItem
trait, so that it can be used for both inputs and outputs.
It simply calls the range_gate method on the composer, putting in the value
of the TransactionItem, and using 2^64 as the upper bound.
This also includes a test for the range gadget.