Finalise gadgets
The gadgets had some minor issues surrounding how they were going/trying to interact with one another. This has been solved in the mod file by commenting out the code until one encompassing gadget has been made. In the `sk_r.rs` the function has been refactored to return empty this allows us to mock this gadget in the circuits. The other changes are the removal of warnings and code cleaning etc.
The only remaining issue is to ensure the methods in `sanity.rs` are coming from the correct imports, as this refactor has lead them to rely upon the BLS.