Compute the value commitment correctly
Previously, the value commitment was simply calculated by
hashing the value and the blinding factor together. In reality,
this should be a Pedersen commitment. This commit should ensure
correct computation of the Pedersen commitment, and should fix
any references to this field (as its type has changed due to the
fix).